Rename opaque.d.ts to tagged.d.ts
I considered leaving the deprecated Opaque types in
opaque.d.ts, but that would require that we export the
shared `Tag` type, and I'm not sure if that would
cause it to be suggested by IDEs.
